Files moved from lib to util
